refactor(entity): remove redundant explicit double conversions in fromJson methods
Simplify numeric field parsing by removing unnecessary `.toDouble()` casts on JSON values
that already expect doubles, relying on the `??` operator to provide the correct type.
This avoids unnecessary explicit conversions and makes the code cleaner.

+++ lib/entity/creature_onkill_reputation_entity.dart
@@ -35,11 +35,9 @@ class CreatureOnKillReputationEntity {
       isTeamAward1: (json['IsTeamAward1'] ?? json['isTeamAward1'] ?? 0) == 1,
       isTeamAward2: (json['IsTeamAward2'] ?? json['isTeamAward2'] ?? 0) == 1,
       rewOnKillRepValue1:
-          (json['RewOnKillRepValue1'] ?? json['rewOnKillRepValue1'] ?? 0.0)
-              .toDouble(),
+          json['RewOnKillRepValue1'] ?? json['rewOnKillRepValue1'] ?? 0.0,
       rewOnKillRepValue2:
-          (json['RewOnKillRepValue2'] ?? json['rewOnKillRepValue2'] ?? 0.0)
-              .toDouble(),
+          json['RewOnKillRepValue2'] ?? json['rewOnKillRepValue2'] ?? 0.0,
       teamDependent: json['TeamDependent'] ?? json['teamDependent'] ?? 0,
     );
   }
